allow playing with others

+++ commands/fun/whosthatpokemon.js
@@ -4,9 +4,11 @@ const Jimp = require('jimp');
 
 module.exports = {
   data: new SlashCommandBuilder().setName('whosthatpokemon')
-  .setDescription('Guess the pokemon'),
+  .setDescription('Guess the pokemon')
+  .addBooleanOption(option => option.setName('private').setDescription('set to false if you want to play with others. Default is true')),
   async execute(interaction) {
-    await interaction.deferReply({ ephemeral: true });
+    const private = interaction.options.getBoolean('private') ?? true;
+    await interaction.deferReply({ ephemeral: private });
     const pokedex = await axios.get('https://pokeapi.co/api/v2/pokemon-species/?limit=0').then(res => res.data); // get the total number of available pokemon to pull from
     const pokedexEntries = pokedex.count;
     const selectedPokemon = Math.floor(Math.random() * pokedexEntries);
@@ -33,27 +35,37 @@ module.exports = {
     const time = 30000; // 30 seconds
     const embed = new EmbedBuilder().setTitle(`Who's that pokemon? ${time / 1000} seconds on the clock!`).setImage(`attachment://${pokemonData.name}Black.png`)
     await interaction.followUp({ embeds: [embed], files: [{ attachment: outputPath, name: `${pokemonData.name}Black.png`}]});
-
-    const collector = interaction.channel.createMessageCollector({
-      filter: (msg) => msg.author.id === interaction.user.id,
+    const collectorObject = {
       max: 1, // Collect only one response
       time: time, // Set a timeout for response collection (30 seconds in this example)
       errors: ['time'] // Handle timeout errors
-    });
+    }
+    if (!private) {
+      collectorObject.filter = (msg) => msg.author.id === interaction.user.id;
+    }
+    const collector = interaction.channel.createMessageCollector(collectorObject);
     
     collector.on('collect', async (msg) => {
-      const userAnswer = msg.content.toLowerCase(); // Convert the answer to lowercase for case insensitivity
-    
-      // Check if the user's answer matches the correct Pokémon name
-      const embed = new EmbedBuilder().setTitle(`The pokemon was ${pokemonData.name}`).setImage(pokemonImageLink);
-      await interaction.followUp({ content: `You are ${userAnswer === pokemonData.name.toLowerCase() ? 'Correct 🎉': 'Incorrect ❌'}`, embeds: [embed], ephemeral: true });
-      await msg.delete().catch();
+      const userAnswer = msg.content.toLowerCase();
+      if (!private && userAnswer === pokemonData.name.toLowerCase()) {
+        collector.stop();
+        const embed = new EmbedBuilder().setTitle(`The pokemon was ${pokemonData.name}`).setImage(pokemonImageLink);
+        await interaction.followUp({ content: `${msg.author.username} is Correct 🎉`, embeds: [embed], ephemeral: !private });
+      } else {
+        collector.stop();
+        const embed = new EmbedBuilder().setTitle(`The pokemon was ${pokemonData.name}`).setImage(pokemonImageLink);
+        await interaction.followUp({ content: `You are ${userAnswer === pokemonData.name.toLowerCase() ? 'Correct 🎉': 'Incorrect ❌'}`, embeds: [embed], ephemeral: private });
+      }
+      
+      if (private) {
+        await msg.delete().catch();
+      }
     });
     
     collector.on('end', (collected, reason) => {
       if (reason === 'time') {
         const embed = new EmbedBuilder().setTitle(`The pokemon was ${pokemonData.name}`).setImage(pokemonImageLink);
-        interaction.followUp({ content: `⌛ Time\'s up! You did not provide an answer.`, embeds: [embed], ephemeral: true });
+        interaction.followUp({ content: `⌛ Time\'s up! You did not provide an answer.`, embeds: [embed], ephemeral: private });
       }
     });
   }
